pyproject.toml
[tool.poetry]
name = "portatal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.9.1"
lxml = "^5.1.0"
uvicorn = "^0.26.0"
fastapi = "^0.109.0"
requests = "^2.31.0"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
